fix(booster): rework booster

+++ src/artemis/simulator/model/material/booster/Booster.java
@@ -3,11 +3,11 @@ package artemis.simulator.model.material.booster;
 public class Booster {
     
     private final String name;
-    private final long thrust;
-    private final long weight;
+    private final double thrust; 
+    private final double weight; 
     private final long price;
 
-    public Booster(String name, long thrust, long weight, long price) {
+    public Booster(String name, double thrust, double weight, long price) {
         this.name = name;
         this.thrust = thrust;
         this.weight = weight;
@@ -15,7 +15,7 @@ public class Booster {
     }
     
     public String getName() { return name; }
-    public long getThrust() { return thrust; }
-    public long getWeight() { return weight; }
+    public double getThrust() { return thrust; }
+    public double getWeight() { return weight; }
     public long getPrice() { return price; }
 }
